Revert "client: Add support for HTTP/HTTPS proxy (#30812)"
This reverts commit d4f47aa653b6016d4158f8f4f3234f97718834ce.
